Move ContextMenuStore to lib/contextMenu
